No URI-encoding for usernames/passwords
URI-encoded usernames and passwords caused url.urlParse to
fail [1], which meant that usernames and passwords with
characters that needed to be URI-encoded couldn't be used.

This patch removes the URI-encoding of usernames and passwords,
and the corresponding decodeURIComponent calls couchdb.js and
repository.js. This does not affect existing functionality as
the auth string is base64-encoded before being added to the
Authorization header.

[1] https://github.com/joyent/node/issues/1163

Conflicts:

	lib/commands/unpublish.js
	lib/commands/upload.js
	src/kanso/commands/listdb.js
	src/kanso/commands/publish.js
	src/kanso/commands/push.js