Add --source-dir option to install command.
Allows to resolve dependencies from local packages without publishing
them. In my case I use multiple packages which all depend on a
common package with some common code and functionality. There's no sense in
publishing it, because it is highly app-specific.

--source-dir option should point to the directory above the local
package(s), similar to --target-dir.

--source-dir supports multiple directories separated by semicolon ';'.
Directories are searched in the order given, left to right.